+++ crates/polars-plan/src/plans/conversion/dsl_to_ir/scans.rs
@@ -129,8 +129,6 @@ pub(super) async fn dsl_to_ir(
             None
         };
 
-        dbg!(&hive_parts);
-
         if let Some(ref hive_parts) = hive_parts {
             let hive_schema = hive_parts.schema();
             file_info.update_schema_with_hive_schema(hive_schema.clone());
